requirements.txt
discord.py>=2.0.0
python-dotenv>=0.19.0
jinja2>=3.0.0
termcolor>=1.1.0 